Almost done, comprobar funcion de error

+++ abb.h
@@ -14,7 +14,7 @@ typedef struct{
     int initVal; //flag para saber si el token esta inicializado
     union{
         float val; //valor del token
-        unsigned (*func)(); //funcion del token
+        double (*func)(); //funcion del token
     }data;
 }tipoelem;
